Client sends data to backend for a query
I made an input box on favdishes.js, added an
event listener to have its contents persistently
be sent to the backend whenever the text is
changed via a POST request using axios.
I then configured the backend to read the request
using Express and and respond back with a json
message dependent on the initial request.
Soon this response (which is currently just
the length of the input string) will be replaced
with an array of meal objects that cointain
information about the meal.