Website restored to previous functionality
All the React/Node.js interactions that were
present with the previous frontend are back.
Only thing of note was one weird issue where
Switch is not present in react-router-dom 6.whate\
ver version we had in the package.json prior.

My next order of business is to enable the user
to select there favorite meals from the one's
displayed on their profile page.
This will entail modularizing the elements on this
page and remembering how state works so that I can
configure each meal's like button so that a click
tells the database the user likes that particular
meal.